ChromaShift/pyproject.toml
[project]
name = "chromashift"
version = "1.0.0"
description = "System-wide color accessibility overlay for Windows colorblind users"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "keyboard>=0.13.5",
]

[project.scripts]
chromashift = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pyinstaller>=6.0",
]
